+++ src/com/aventura/math/vector/TestMatrix3.java
@@ -11,8 +11,8 @@ public class TestMatrix3 {
 		System.out.println("***** Test Matrix3 : testMatrix3_0 *****");
 		
 		Matrix3 A = new Matrix3();
-		Matrix3 B = new Matrix3(0.0);
-		Matrix3 C = new Matrix3(7.0);
+		Matrix3 B = new Matrix3(0);
+		Matrix3 C = new Matrix3(7);
 		System.out.println("A="+A);
 		System.out.println("B="+B);
 		System.out.println("C="+C);
@@ -24,7 +24,7 @@ public class TestMatrix3 {
 	public void testMatrix3_array_0() {
 		System.out.println("***** Test Matrix3 : testMatrix3_array() *****");
 		
-		double[][] array = new double[3][3];
+		float[][] array = new float[3][3];
 		
 		for (int i=0; i<3; i++) {
 			for (int j=0; j<3; j++) {
@@ -51,21 +51,21 @@ public class TestMatrix3 {
 	public void testMatrix3_array_value() {
 		System.out.println("***** Test Matrix3 : testMatrix3_array_value() *****");
 		
-		double[][] array = new double[3][3];
+		float[][] array = new float[3][3];
 		
 		for (int i=0; i<3; i++) {
 			for (int j=0; j<3; j++) {
 				array[i][j] = 5;
 			}
 		}
-		array[1][2]=22.3;
+		array[1][2]=22.3f;
 
 		Matrix3 A;
 		Matrix3 B;
 		try {
 			A = new Matrix3(array);
-			B = new Matrix3(5.0);
-			B.set(1,2,22.3);
+			B = new Matrix3(5);
+			B.set(1,2,22.3f);
 			System.out.println("A="+A);
 			System.out.println("B="+B);
 	
@@ -81,7 +81,7 @@ public class TestMatrix3 {
 	public void testMatrix3_plus() {
 		System.out.println("***** Test Matrix3 : testMatrix3_plus() *****");
 		
-		double[][] array = new double[3][3];
+		float[][] array = new float[3][3];
 		
 		for (int i=0; i<3; i++) {
 			for (int j=0; j<3; j++) {
@@ -94,7 +94,7 @@ public class TestMatrix3 {
 		Matrix3 C;
 		try {
 			A = new Matrix3(array);
-			B = new Matrix3(5.0);
+			B = new Matrix3(5);
 			C = A.plus(B);
 			System.out.println("A="+A);
 			System.out.println("B="+B);
@@ -114,7 +114,7 @@ public class TestMatrix3 {
 	public void testMatrix3_minus() {
 		System.out.println("***** Test Matrix3 : testMatrix3_minus() *****");
 		
-		double[][] array = new double[3][3];
+		float[][] array = new float[3][3];
 		
 		for (int i=0; i<3; i++) {
 			for (int j=0; j<3; j++) {
@@ -127,7 +127,7 @@ public class TestMatrix3 {
 		Matrix3 C;
 		try {
 			A = new Matrix3(array);
-			B = new Matrix3(2.0);
+			B = new Matrix3(2);
 			C = A.minus(B);
 			
 			System.out.println("A="+A);
@@ -148,8 +148,8 @@ public class TestMatrix3 {
 	public void testMatrix3_plusEquals() {
 		System.out.println("***** Test Matrix3 : testMatrix3_plusEquals() *****");
 		
-		double[][] array1 = new double[3][3];
-		double[][] array2 = new double[3][3];
+		float[][] array1 = new float[3][3];
+		float[][] array2 = new float[3][3];
 		
 		for (int i=0; i<3; i++) {
 			for (int j=0; j<3; j++) {
@@ -191,8 +191,8 @@ public class TestMatrix3 {
 	public void testMatrix3_minusEquals() {
 		System.out.println("***** Test Matrix3 : testMatrix3_minusEquals() *****");
 		
-		double[][] array1 = new double[3][3];
-		double[][] array2 = new double[3][3];
+		float[][] array1 = new float[3][3];
+		float[][] array2 = new float[3][3];
 		
 		for (int i=0; i<3; i++) {
 			for (int j=0; j<3; j++) {
@@ -234,8 +234,8 @@ public class TestMatrix3 {
 	public void testMatrix3_times() {
 		System.out.println("***** Test Matrix3 : testMatrix3_times() *****");
 		
-		double[][] array1 = new double[3][3];
-		double[][] array2 = new double[3][3];
+		float[][] array1 = new float[3][3];
+		float[][] array2 = new float[3][3];
 		
 		for (int i=0; i<3; i++) {
 			for (int j=0; j<3; j++) {
